add image_reshapers.py
